Time: 3 ms (65.54%), Space: 6.2 MB (8.19%) - LeetHub

+++ 0509-fibonacci-number/0509-fibonacci-number.cpp
@@ -1,6 +1,6 @@
 class Solution {
-    const double phi = (sqrt(5) + 1) / 2; 
-    const double sqrt5 = sqrt(5);
+    const float phi = (sqrt(5) + 1) / 2; 
+    const float sqrt5 = sqrt(5);
 public:
     int fib(int n) {
         return round(pow(phi, n) / sqrt5);
